enable pch generation mode for utests
basically we are back to a functional plex.exe that we can use
automatically for unit tests.

Removed /gen/ vestiges.

+++ src/utests/stdafx.h
@@ -813,7 +813,7 @@ public:
     return (handle_ != INVALID_HANDLE_VALUE);
   }
 
-  size_t size_in_bytes() const {
+  long long size_in_bytes() const {
     LARGE_INTEGER li = {0};
     ::GetFileSizeEx(handle_, &li);
     return li.QuadPart;
